Add wasm32 workaround for missing `pipelined_rendering`
The `pipelined_rendering` export is no longer available in Bevy since
version 0.12 in wasm32 builds; it can no longer be used to select the
right subapp.

Add a dummy version of the `AppLabel` that targets the
`pipelined_rendering` subapp. Using a dummy app label allows to keep the
same code to get the right subapp to apply the framepace resources and
systems.